Refactor XMLA and DAX execution internals
