+++ src/include/dax_probe.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <string>
+
+namespace duckdb {
+
+std::string BuildLimitedDaxSchemaProbe(const std::string &statement,
+                                       int64_t row_limit);
+std::string BuildDaxSchemaProbeForTesting(const std::string &statement,
+                                          int64_t row_limit);
+
+} // namespace duckdb
