+++ src/include/pbi_scanner_test_functions.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace duckdb {
+
+class ExtensionLoader;
+
+void RegisterPbiScannerTestFunctions(ExtensionLoader &loader);
+
+} // namespace duckdb
