+++ src/include/pbi_scanner_util.hpp
@@ -7,6 +7,7 @@
 #include <cstdint>
 #include <cstdio>
 #include <cstdlib>
+#include <string>
 
 namespace duckdb {
 
@@ -31,6 +32,25 @@ inline void DebugTiming(const char *label,
                         .count();
   std::fprintf(stderr, "[pbi_scanner] %s: %lld ms\n", label,
                static_cast<long long>(elapsed_ms));
+}
+
+inline int64_t CurrentUnixSeconds() {
+  return std::chrono::duration_cast<std::chrono::seconds>(
+             std::chrono::system_clock::now().time_since_epoch())
+      .count();
+}
+
+inline uint64_t Fnv1a64(const string &value) {
+  uint64_t hash = 1469598103934665603ULL;
+  for (auto ch : value) {
+    hash ^= static_cast<uint8_t>(ch);
+    hash *= 1099511628211ULL;
+  }
+  return hash;
+}
+
+inline string HashSensitiveValue(const string &value) {
+  return std::to_string(Fnv1a64(value));
 }
 
 inline uint8_t DecodeHexDigit(char value, const char *error_message) {
