+++ src/include/xmla_util.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <string>
+
+namespace duckdb {
+
+std::string EscapeXML(const std::string &value);
+std::string UnescapeXML(const std::string &value);
+std::string DecodeXMLName(const std::string &raw);
+std::string ExtractLocalName(const std::string &name);
+void TrimString(std::string &value);
+
+} // namespace duckdb
