+++ src/powerbi_resolver.cpp
@@ -8,7 +8,6 @@
 
 #include "yyjson.hpp"
 
-#include <chrono>
 #include <cstdio>
 #include <cstdlib>
 #include <mutex>
@@ -41,21 +40,6 @@ struct CachedMwcToken {
 
 std::mutex mwc_token_cache_lock;
 std::unordered_map<string, CachedMwcToken> mwc_token_cache;
-
-static int64_t CurrentUnixSeconds() {
-  return std::chrono::duration_cast<std::chrono::seconds>(
-             std::chrono::system_clock::now().time_since_epoch())
-      .count();
-}
-
-static string HashSensitiveValue(const string &value) {
-  uint64_t hash = 1469598103934665603ULL;
-  for (auto ch : value) {
-    hash ^= static_cast<uint8_t>(ch);
-    hash *= 1099511628211ULL;
-  }
-  return std::to_string(hash);
-}
 
 static HttpHeaders PowerBIHeaders(const string &access_token) {
   return HttpHeaders{std::make_pair("Authorization", "Bearer " + access_token)};
