+++ src/xmla.cpp
@@ -1,21 +1,21 @@
 #include "xmla.hpp"
 #include "pbi_scanner_util.hpp"
+#include "xmla_transport.hpp"
+#include "xmla_util.hpp"
+#include "xmla_value.hpp"
+#include "xpress.hpp"
 
 #include "duckdb/common/case_insensitive_map.hpp"
 #include "duckdb/common/exception.hpp"
-#include "duckdb/common/operator/double_cast_operator.hpp"
 #include "duckdb/common/string_util.hpp"
 #include "duckdb/common/types/date.hpp"
 #include "duckdb/common/types/time.hpp"
 #include "duckdb/common/types/timestamp.hpp"
 
-#include <cerrno>
 #include <chrono>
 #include <algorithm>
-#include <cmath>
 #include <condition_variable>
 #include <cstdio>
-#include <cstdlib>
 #include <cstring>
 #include <deque>
 #include <iomanip>
@@ -29,83 +29,6 @@ namespace duckdb {
 
 namespace {
 
-static std::string EscapeXML(const std::string &value) {
-  std::string escaped;
-  escaped.reserve(value.size() + 16);
-  for (auto ch : value) {
-    switch (ch) {
-    case '&':
-      escaped += "&amp;";
-      break;
-    case '<':
-      escaped += "&lt;";
-      break;
-    case '>':
-      escaped += "&gt;";
-      break;
-    case '"':
-      escaped += "&quot;";
-      break;
-    case '\'':
-      escaped += "&apos;";
-      break;
-    default:
-      escaped.push_back(ch);
-      break;
-    }
-  }
-  return escaped;
-}
-
-enum class XmlaTransportMode : uint8_t { PLAIN, SX, XPRESS, SX_XPRESS };
-
-static XmlaTransportMode ResolveXmlaTransportMode() {
-  auto *raw_value = std::getenv("PBI_SCANNER_XMLA_TRANSPORT");
-  if (!raw_value || !*raw_value) {
-    return XmlaTransportMode::SX_XPRESS;
-  }
-  auto value = StringUtil::Lower(Trimmed(raw_value));
-  if (value == "plain") {
-    return XmlaTransportMode::PLAIN;
-  }
-  if (value == "sx") {
-    return XmlaTransportMode::SX;
-  }
-  if (value == "xpress") {
-    return XmlaTransportMode::XPRESS;
-  }
-  if (value == "sx_xpress" || value == "sxxpress" || value == "sx+xpress") {
-    return XmlaTransportMode::SX_XPRESS;
-  }
-  throw InvalidInputException(
-      "PBI_SCANNER_XMLA_TRANSPORT must be plain, sx, xpress, or sx_xpress");
-}
-
-static const char *XmlaTransportFlags(XmlaTransportMode mode) {
-  switch (mode) {
-  case XmlaTransportMode::PLAIN:
-    return "0,0,0,0,0";
-  case XmlaTransportMode::SX:
-    return "1,0,0,1,0";
-  case XmlaTransportMode::XPRESS:
-    return "1,0,0,0,1";
-  case XmlaTransportMode::SX_XPRESS:
-    return "1,0,0,1,1";
-  }
-  return "0,0,0,0,0";
-}
-
-static XmlaTransportMode SchemaProbeTransportMode(XmlaTransportMode mode) {
-  if (mode == XmlaTransportMode::SX || mode == XmlaTransportMode::SX_XPRESS) {
-    return XmlaTransportMode::XPRESS;
-  }
-  return mode;
-}
-
-static bool XmlaTransportNeedsProtocolCapabilities(XmlaTransportMode mode) {
-  return mode != XmlaTransportMode::PLAIN;
-}
-
 static void AppendProtocolCapabilities(std::string &envelope,
                                        XmlaTransportMode mode) {
   if (!XmlaTransportNeedsProtocolCapabilities(mode)) {
@@ -114,7 +37,7 @@ static void AppendProtocolCapabilities(std::string &envelope,
   envelope += "<ProtocolCapabilities "
               "xmlns=\"http://schemas.microsoft.com/analysisservices/2003/"
               "engine\">";
-  if (mode == XmlaTransportMode::SX || mode == XmlaTransportMode::SX_XPRESS) {
+  if (XmlaTransportIsSx(mode)) {
     envelope += "<Capability>sx</Capability>";
   }
   if (mode == XmlaTransportMode::XPRESS ||
@@ -124,794 +47,10 @@ static void AppendProtocolCapabilities(std::string &envelope,
   envelope += "</ProtocolCapabilities>";
 }
 
-static std::string UnescapeXML(const std::string &value) {
-  if (value.find('&') == std::string::npos) {
-    return value;
-  }
-  std::string result;
-  result.reserve(value.size());
-  for (idx_t i = 0; i < value.size(); i++) {
-    if (value[i] != '&') {
-      result.push_back(value[i]);
-      continue;
-    }
-    auto semi = value.find(';', i + 1);
-    if (semi == std::string::npos) {
-      result.push_back(value[i]);
-      continue;
-    }
-    auto entity = value.substr(i + 1, semi - i - 1);
-    if (entity == "amp") {
-      result.push_back('&');
-    } else if (entity == "lt") {
-      result.push_back('<');
-    } else if (entity == "gt") {
-      result.push_back('>');
-    } else if (entity == "quot") {
-      result.push_back('"');
-    } else if (entity == "apos") {
-      result.push_back('\'');
-    } else if (!entity.empty() && entity[0] == '#') {
-      char *end_ptr = nullptr;
-      long codepoint = 0;
-      if (entity.size() > 2 && (entity[1] == 'x' || entity[1] == 'X')) {
-        codepoint = std::strtol(entity.c_str() + 2, &end_ptr, 16);
-      } else {
-        codepoint = std::strtol(entity.c_str() + 1, &end_ptr, 10);
-      }
-      if (end_ptr && *end_ptr == '\0' && codepoint >= 0 && codepoint <= 0x7F) {
-        result.push_back(static_cast<char>(codepoint));
-      } else {
-        result.append(value, i, semi - i + 1);
-      }
-    } else {
-      result.append(value, i, semi - i + 1);
-    }
-    i = semi;
-  }
-  return result;
-}
-
-static std::string DecodeXMLName(const std::string &raw) {
-  if (raw.find("_x") == std::string::npos) {
-    return raw;
-  }
-
-  std::string result;
-  result.reserve(raw.size());
-  for (idx_t i = 0; i < raw.size();) {
-    if (i + 7 <= raw.size() && raw[i] == '_' &&
-        (raw[i + 1] == 'x' || raw[i + 1] == 'X') && raw[i + 6] == '_') {
-      auto hex = raw.substr(i + 2, 4);
-      char *end_ptr = nullptr;
-      auto codepoint = std::strtol(hex.c_str(), &end_ptr, 16);
-      if (end_ptr && *end_ptr == '\0' && codepoint >= 0 && codepoint <= 0x7F) {
-        result.push_back(static_cast<char>(codepoint));
-        i += 7;
-        continue;
-      }
-    }
-    result.push_back(raw[i++]);
-  }
-  return result;
-}
-
-static std::string ExtractLocalName(const std::string &name) {
-  auto colon = name.find(':');
-  return colon == std::string::npos ? name : name.substr(colon + 1);
-}
-
-static void TrimString(std::string &value) {
-  idx_t begin = 0;
-  while (
-      begin < value.size() &&
-      StringUtil::CharacterIsSpace(static_cast<unsigned char>(value[begin]))) {
-    begin++;
-  }
-  idx_t end = value.size();
-  while (end > begin && StringUtil::CharacterIsSpace(
-                            static_cast<unsigned char>(value[end - 1]))) {
-    end--;
-  }
-  if (begin == 0 && end == value.size()) {
-    return;
-  }
-  value = value.substr(begin, end - begin);
-}
-
 static bool TagStartsWith(const std::string &tag, const char *prefix,
                           size_t prefix_len) {
   return tag.size() >= prefix_len &&
          std::memcmp(tag.data(), prefix, prefix_len) == 0;
-}
-
-static uint16_t ReadLittleEndian16(const_data_ptr_t data, idx_t size,
-                                   idx_t offset) {
-  if (offset + 2 > size) {
-    throw IOException("XPRESS payload ended unexpectedly");
-  }
-  return static_cast<uint16_t>(data[offset]) |
-         static_cast<uint16_t>(data[offset + 1] << 8);
-}
-
-static uint32_t ReadLittleEndian32(const_data_ptr_t data, idx_t size,
-                                   idx_t offset) {
-  if (offset + 4 > size) {
-    throw IOException("XPRESS payload ended unexpectedly");
-  }
-  return static_cast<uint32_t>(data[offset]) |
-         (static_cast<uint32_t>(data[offset + 1]) << 8) |
-         (static_cast<uint32_t>(data[offset + 2]) << 16) |
-         (static_cast<uint32_t>(data[offset + 3]) << 24);
-}
-
-static void DecompressXpressLz77Block(const_data_ptr_t input, idx_t input_size,
-                                      std::string &output,
-                                      idx_t expected_output_size) {
-  idx_t input_pos = 0;
-  idx_t output_start = output.size();
-  idx_t last_length_half_byte = DConstants::INVALID_INDEX;
-
-  while (output.size() - output_start < expected_output_size) {
-    auto flags = ReadLittleEndian32(input, input_size, input_pos);
-    input_pos += 4;
-
-    for (int bit = 31;
-         bit >= 0 && output.size() - output_start < expected_output_size;
-         bit--) {
-      if ((flags & (uint32_t(1) << bit)) == 0) {
-        if (input_pos >= input_size) {
-          throw IOException("XPRESS literal exceeded compressed block");
-        }
-        output.push_back(static_cast<char>(input[input_pos++]));
-        continue;
-      }
-
-      auto match_bytes = ReadLittleEndian16(input, input_size, input_pos);
-      input_pos += 2;
-      uint32_t match_length = match_bytes & 0x7;
-      auto match_offset = static_cast<idx_t>((match_bytes >> 3) + 1);
-      if (match_length == 7) {
-        if (last_length_half_byte == DConstants::INVALID_INDEX) {
-          if (input_pos >= input_size) {
-            throw IOException("XPRESS match length exceeded compressed block");
-          }
-          match_length = input[input_pos] & 0x0F;
-          last_length_half_byte = input_pos++;
-        } else {
-          match_length = input[last_length_half_byte] >> 4;
-          last_length_half_byte = DConstants::INVALID_INDEX;
-        }
-        if (match_length == 15) {
-          if (input_pos >= input_size) {
-            throw IOException("XPRESS extended match length missing");
-          }
-          match_length = input[input_pos++];
-          if (match_length == 255) {
-            match_length = ReadLittleEndian16(input, input_size, input_pos);
-            input_pos += 2;
-            if (match_length == 0) {
-              match_length = ReadLittleEndian32(input, input_size, input_pos);
-              input_pos += 4;
-            }
-            if (match_length < 22) {
-              throw IOException("XPRESS extended match length was invalid");
-            }
-            match_length -= 22;
-          }
-          match_length += 15;
-        }
-        match_length += 7;
-      }
-      match_length += 3;
-
-      if (match_offset == 0 || match_offset > output.size() - output_start) {
-        throw IOException("XPRESS match offset was invalid");
-      }
-      for (uint32_t i = 0; i < match_length; i++) {
-        output.push_back(output[output.size() - match_offset]);
-        if (output.size() - output_start > expected_output_size) {
-          throw IOException("XPRESS match exceeded expected block size");
-        }
-      }
-    }
-  }
-}
-
-static std::string DecompressXpressLz77Framed(const std::string &compressed) {
-  std::string output;
-  auto input = const_data_ptr_cast(compressed.data());
-  auto input_size = static_cast<idx_t>(compressed.size());
-  idx_t input_pos = 0;
-
-  while (input_pos < input_size) {
-    auto original_size = ReadLittleEndian32(input, input_size, input_pos);
-    input_pos += 4;
-    auto compressed_size = ReadLittleEndian32(input, input_size, input_pos);
-    input_pos += 4;
-    if (original_size > 65535 || compressed_size > 65535) {
-      throw IOException("XPRESS block sizes must not exceed 65535 bytes");
-    }
-    if (input_pos + compressed_size > input_size) {
-      throw IOException("XPRESS block exceeded payload size");
-    }
-    if (original_size == compressed_size) {
-      output.append(reinterpret_cast<const char *>(input + input_pos),
-                    compressed_size);
-    } else {
-      DecompressXpressLz77Block(input + input_pos, compressed_size, output,
-                                original_size);
-    }
-    input_pos += compressed_size;
-  }
-  return output;
-}
-
-class XpressLz77FrameStream {
-public:
-  bool
-  Feed(const_data_ptr_t data, idx_t size,
-       const std::function<bool(const_data_ptr_t, idx_t)> &on_decompressed) {
-    buffer.append(reinterpret_cast<const char *>(data), size);
-    idx_t offset = 0;
-    while (buffer.size() - offset >= FRAME_HEADER_SIZE) {
-      auto input = const_data_ptr_cast(buffer.data());
-      auto original_size = ReadLittleEndian32(input, buffer.size(), offset);
-      auto compressed_size =
-          ReadLittleEndian32(input, buffer.size(), offset + 4);
-      if (original_size > 65535 || compressed_size > 65535) {
-        throw IOException("XPRESS block sizes must not exceed 65535 bytes");
-      }
-      auto frame_size = FRAME_HEADER_SIZE + static_cast<idx_t>(compressed_size);
-      if (buffer.size() - offset < frame_size) {
-        break;
-      }
-
-      std::string output;
-      output.reserve(original_size);
-      auto frame_payload = input + offset + FRAME_HEADER_SIZE;
-      if (original_size == compressed_size) {
-        output.append(reinterpret_cast<const char *>(frame_payload),
-                      compressed_size);
-      } else {
-        DecompressXpressLz77Block(frame_payload, compressed_size, output,
-                                  original_size);
-      }
-      offset += frame_size;
-      decompressed_bytes += output.size();
-      if (!output.empty() &&
-          !on_decompressed(const_data_ptr_cast(output.data()), output.size())) {
-        Compact(offset);
-        return false;
-      }
-    }
-    Compact(offset);
-    return true;
-  }
-
-  void Finish() const {
-    if (!buffer.empty()) {
-      throw IOException("XPRESS payload ended unexpectedly");
-    }
-  }
-
-  idx_t BufferedBytes() const { return buffer.size(); }
-  idx_t DecompressedBytes() const { return decompressed_bytes; }
-
-private:
-  static constexpr idx_t FRAME_HEADER_SIZE = 8;
-
-  void Compact(idx_t offset) {
-    if (offset == 0) {
-      return;
-    }
-    buffer.erase(0, offset);
-  }
-
-  std::string buffer;
-  idx_t decompressed_bytes = 0;
-};
-
-static bool TryParseInt64(const std::string &value, int64_t &result) {
-  if (value.empty()) {
-    return false;
-  }
-  constexpr size_t kStackMax = 128;
-  char stack[kStackMax];
-  const char *cstr;
-  std::string heap;
-  if (value.size() < kStackMax) {
-    std::memcpy(stack, value.data(), value.size());
-    stack[value.size()] = '\0';
-    cstr = stack;
-  } else {
-    heap = value;
-    cstr = heap.c_str();
-  }
-  errno = 0;
-  char *end_ptr = nullptr;
-  auto parsed = std::strtoll(cstr, &end_ptr, 10);
-  if (errno != 0 || !end_ptr) {
-    return false;
-  }
-  if (static_cast<size_t>(end_ptr - cstr) != value.size()) {
-    return false;
-  }
-  result = parsed;
-  return true;
-}
-
-static bool TryParseUInt64(const std::string &value, uint64_t &result) {
-  if (value.empty()) {
-    return false;
-  }
-  constexpr size_t kStackMax = 128;
-  char stack[kStackMax];
-  const char *cstr;
-  std::string heap;
-  if (value.size() < kStackMax) {
-    std::memcpy(stack, value.data(), value.size());
-    stack[value.size()] = '\0';
-    cstr = stack;
-  } else {
-    heap = value;
-    cstr = heap.c_str();
-  }
-  errno = 0;
-  char *end_ptr = nullptr;
-  auto parsed = std::strtoull(cstr, &end_ptr, 10);
-  if (errno != 0 || !end_ptr) {
-    return false;
-  }
-  if (static_cast<size_t>(end_ptr - cstr) != value.size()) {
-    return false;
-  }
-  result = parsed;
-  return true;
-}
-
-static bool TryParseDouble(const std::string &value, double &result) {
-  return TryDoubleCast(value.data(), static_cast<idx_t>(value.size()), result,
-                       true);
-}
-
-static bool HasTimezoneOffset(const std::string &value) {
-  if (value.empty()) {
-    return false;
-  }
-  auto last_char = value.back();
-  if (last_char == 'Z' || last_char == 'z') {
-    return true;
-  }
-  auto separator = value.find('T');
-  if (separator == std::string::npos) {
-    separator = value.find(' ');
-  }
-  if (separator == std::string::npos) {
-    return false;
-  }
-  for (idx_t i = separator + 1; i < value.size(); i++) {
-    if ((value[i] == '+' || value[i] == '-') && i + 2 < value.size()) {
-      return true;
-    }
-  }
-  return false;
-}
-
-static bool TryParseDateValue(const std::string &value, Value &result) {
-  if (value.find('T') != std::string::npos ||
-      value.find(' ') != std::string::npos) {
-    return false;
-  }
-  try {
-    result = Value::DATE(Date::FromString(value, true));
-    return true;
-  } catch (const Exception &) {
-    return false;
-  } catch (const std::exception &) {
-    return false;
-  }
-}
-
-static bool TryParseTimeValue(const std::string &value, Value &result) {
-  if (value.find('T') != std::string::npos ||
-      value.find(' ') != std::string::npos) {
-    return false;
-  }
-  try {
-    result = Value::TIME(Time::FromString(value, true));
-    return true;
-  } catch (const Exception &) {
-    return false;
-  } catch (const std::exception &) {
-    return false;
-  }
-}
-
-static bool TryParseTimestampValue(const std::string &value, Value &result) {
-  try {
-    auto has_offset = HasTimezoneOffset(value);
-    auto timestamp = Timestamp::FromString(value, has_offset);
-    if (has_offset) {
-      result = Value::TIMESTAMPTZ(timestamp_tz_t(timestamp));
-    } else {
-      result = Value::TIMESTAMP(timestamp);
-    }
-    return true;
-  } catch (const Exception &) {
-    return false;
-  } catch (const std::exception &) {
-    return false;
-  }
-}
-
-static bool TryDecodeSsasTemporalSerial(double serial_value, date_t &out_date,
-                                        dtime_t &out_time) {
-  if (!std::isfinite(serial_value)) {
-    return false;
-  }
-  // SSAS commonly uses OLE Automation serial date semantics:
-  // days since 1899-12-30, including fractional day for time.
-  static constexpr int64_t MICROS_PER_DAY = 24LL * 60LL * 60LL * 1000000LL;
-  static const int32_t OLE_BASE_EPOCH_DAYS =
-      Date::EpochDays(Date::FromString("1899-12-30", true));
-
-  auto total_micros =
-      static_cast<int64_t>(serial_value * static_cast<double>(MICROS_PER_DAY) +
-                           (serial_value >= 0 ? 0.5 : -0.5));
-  auto day_delta = total_micros / MICROS_PER_DAY;
-  auto micros_in_day = total_micros % MICROS_PER_DAY;
-  if (micros_in_day < 0) {
-    micros_in_day += MICROS_PER_DAY;
-    day_delta--;
-  }
-  auto date_days = static_cast<int64_t>(OLE_BASE_EPOCH_DAYS) + day_delta;
-  if (date_days < NumericLimits<int32_t>::Minimum() ||
-      date_days > NumericLimits<int32_t>::Maximum()) {
-    return false;
-  }
-  out_date = Date::EpochDaysToDate(static_cast<int32_t>(date_days));
-  out_time = dtime_t(micros_in_day);
-  return true;
-}
-
-static bool TryParseTimeFromBaseDateTimestamp(const std::string &value,
-                                              Value &result) {
-  if (HasTimezoneOffset(value)) {
-    return false;
-  }
-  try {
-    auto timestamp = Timestamp::FromString(value, false);
-    auto date = Timestamp::GetDate(timestamp);
-    if (Date::ToString(date) != "1899-12-30") {
-      return false;
-    }
-    result = Value::TIME(Timestamp::GetTime(timestamp));
-    return true;
-  } catch (const Exception &) {
-    return false;
-  } catch (const std::exception &) {
-    return false;
-  }
-}
-
-static bool TryParseDateFromZeroTimeTimestamp(const std::string &value,
-                                              Value &result) {
-  if (value.size() < 19 || (value[10] != 'T' && value[10] != ' ')) {
-    return false;
-  }
-  if (HasTimezoneOffset(value)) {
-    return false;
-  }
-  if (value.compare(11, 8, "00:00:00") != 0) {
-    return false;
-  }
-  return TryParseDateValue(value.substr(0, 10), result);
-}
-
-static std::string NormalizeXmlType(const std::string &source_type) {
-  return StringUtil::Lower(Trimmed(source_type));
-}
-
-static LogicalType MapXmlTypeToLogicalType(const std::string &source_type) {
-  auto normalized = NormalizeXmlType(source_type);
-  if (normalized == "xsd:boolean" || normalized == "boolean") {
-    return LogicalType::BOOLEAN;
-  }
-  if (normalized == "xsd:byte" || normalized == "xsd:short" ||
-      normalized == "xsd:int" || normalized == "xsd:integer" ||
-      normalized == "xsd:long" || normalized == "byte" ||
-      normalized == "short" || normalized == "int" || normalized == "integer" ||
-      normalized == "long") {
-    return LogicalType::BIGINT;
-  }
-  if (normalized == "xsd:unsignedbyte" || normalized == "xsd:unsignedshort" ||
-      normalized == "xsd:unsignedint" || normalized == "xsd:unsignedlong" ||
-      normalized == "unsignedbyte" || normalized == "unsignedshort" ||
-      normalized == "unsignedint" || normalized == "unsignedlong") {
-    return LogicalType::UBIGINT;
-  }
-  if (normalized == "xsd:decimal" || normalized == "xsd:double" ||
-      normalized == "xsd:float" || normalized == "decimal" ||
-      normalized == "double" || normalized == "float") {
-    return LogicalType::DOUBLE;
-  }
-  if (normalized == "xsd:date" || normalized == "date") {
-    return LogicalType::DATE;
-  }
-  if (normalized == "xsd:time" || normalized == "time") {
-    return LogicalType::TIME;
-  }
-  if (normalized == "xsd:datetimeoffset" || normalized == "datetimeoffset") {
-    return LogicalType::TIMESTAMP_TZ;
-  }
-  if (normalized == "xsd:datetime" || normalized == "datetime") {
-    return LogicalType::TIMESTAMP;
-  }
-  return LogicalType::VARCHAR;
-}
-
-static XmlaCoercionKind CoercionKindFromLogicalType(const LogicalType &type) {
-  switch (type.id()) {
-  case LogicalTypeId::BOOLEAN:
-    return XmlaCoercionKind::BOOLEAN;
-  case LogicalTypeId::BIGINT:
-  case LogicalTypeId::HUGEINT:
-    return XmlaCoercionKind::BIGINT;
-  case LogicalTypeId::UBIGINT:
-    return XmlaCoercionKind::UBIGINT;
-  case LogicalTypeId::DOUBLE:
-    return XmlaCoercionKind::DOUBLE;
-  case LogicalTypeId::DATE:
-    return XmlaCoercionKind::DATE;
-  case LogicalTypeId::TIME:
-    return XmlaCoercionKind::TIME;
-  case LogicalTypeId::TIMESTAMP:
-    return XmlaCoercionKind::TIMESTAMP;
-  case LogicalTypeId::TIMESTAMP_TZ:
-    return XmlaCoercionKind::TIMESTAMP_TZ;
-  case LogicalTypeId::SQLNULL:
-    return XmlaCoercionKind::INFER;
-  default:
-    return XmlaCoercionKind::VARCHAR;
-  }
-}
-
-static XmlaCoercionKind
-CoercionKindFromXmlType(const std::string &source_type) {
-  return CoercionKindFromLogicalType(MapXmlTypeToLogicalType(source_type));
-}
-
-static Value InferUntypedXmlValue(const std::string &body) {
-  if (StringUtil::CIEquals(body.data(), body.size(), "true", 4)) {
-    return Value::BOOLEAN(true);
-  }
-  if (StringUtil::CIEquals(body.data(), body.size(), "false", 5)) {
-    return Value::BOOLEAN(false);
-  }
-
-  int64_t signed_value;
-  if (TryParseInt64(body, signed_value)) {
-    return Value::BIGINT(signed_value);
-  }
-
-  uint64_t unsigned_value;
-  if (TryParseUInt64(body, unsigned_value)) {
-    return Value::UBIGINT(unsigned_value);
-  }
-
-  double double_value;
-  if (TryParseDouble(body, double_value)) {
-    return Value::DOUBLE(double_value);
-  }
-
-  std::string temporal_source(body);
-  Value temporal_value;
-  if (TryParseDateValue(temporal_source, temporal_value)) {
-    return temporal_value;
-  }
-  if (TryParseTimeValue(temporal_source, temporal_value)) {
-    return temporal_value;
-  }
-  if (TryParseTimeFromBaseDateTimestamp(temporal_source, temporal_value)) {
-    return temporal_value;
-  }
-  if (TryParseDateFromZeroTimeTimestamp(temporal_source, temporal_value)) {
-    return temporal_value;
-  }
-  if (TryParseTimestampValue(temporal_source, temporal_value)) {
-    return temporal_value;
-  }
-
-  return Value(body);
-}
-
-static Value CoerceXmlValue(const std::string &raw_value,
-                            XmlaCoercionKind coercion_kind) {
-  std::string body = raw_value;
-  if (raw_value.find('&') != std::string::npos) {
-    body = UnescapeXML(raw_value);
-  }
-  TrimString(body);
-  if (body.empty()) {
-    if (coercion_kind == XmlaCoercionKind::VARCHAR) {
-      return Value(std::string());
-    }
-    if (coercion_kind == XmlaCoercionKind::BOOLEAN) {
-      // SSAS binary metadata frequently encodes false booleans as empty text.
-      return Value::BOOLEAN(false);
-    }
-    return Value();
-  }
-
-  if (coercion_kind == XmlaCoercionKind::INFER) {
-    return InferUntypedXmlValue(body);
-  }
-  if (coercion_kind == XmlaCoercionKind::VARCHAR) {
-    return Value(body);
-  }
-  if (coercion_kind == XmlaCoercionKind::BOOLEAN) {
-    if (StringUtil::CIEquals(body.data(), body.size(), "true", 4) ||
-        (body.size() == 1 && body[0] == '1')) {
-      return Value::BOOLEAN(true);
-    }
-    if (StringUtil::CIEquals(body.data(), body.size(), "false", 5) ||
-        (body.size() == 1 && body[0] == '0')) {
-      return Value::BOOLEAN(false);
-    }
-  }
-  if (coercion_kind == XmlaCoercionKind::BIGINT) {
-    int64_t parsed = 0;
-    if (TryParseInt64(body, parsed)) {
-      return Value::BIGINT(parsed);
-    }
-  }
-  if (coercion_kind == XmlaCoercionKind::UBIGINT) {
-    uint64_t parsed = 0;
-    if (TryParseUInt64(body, parsed)) {
-      return Value::UBIGINT(parsed);
-    }
-  }
-  if (coercion_kind == XmlaCoercionKind::DOUBLE) {
-    double parsed = 0;
-    if (TryParseDouble(body, parsed)) {
-      return Value::DOUBLE(parsed);
-    }
-  }
-
-  if (coercion_kind == XmlaCoercionKind::DATE ||
-      coercion_kind == XmlaCoercionKind::TIME ||
-      coercion_kind == XmlaCoercionKind::TIMESTAMP ||
-      coercion_kind == XmlaCoercionKind::TIMESTAMP_TZ) {
-    double serial_value = 0;
-    if (TryParseDouble(body, serial_value)) {
-      date_t parsed_date;
-      dtime_t parsed_time;
-      if (TryDecodeSsasTemporalSerial(serial_value, parsed_date, parsed_time)) {
-        if (coercion_kind == XmlaCoercionKind::DATE) {
-          return Value::DATE(parsed_date);
-        }
-        if (coercion_kind == XmlaCoercionKind::TIME) {
-          return Value::TIME(parsed_time);
-        }
-        auto parsed_timestamp =
-            Timestamp::FromDatetime(parsed_date, parsed_time);
-        if (coercion_kind == XmlaCoercionKind::TIMESTAMP_TZ) {
-          return Value::TIMESTAMPTZ(timestamp_tz_t(parsed_timestamp));
-        }
-        return Value::TIMESTAMP(parsed_timestamp);
-      }
-    }
-  }
-
-  std::string temporal_source(body);
-  Value temporal_value;
-  if (coercion_kind == XmlaCoercionKind::DATE &&
-      (TryParseDateValue(temporal_source, temporal_value) ||
-       TryParseDateFromZeroTimeTimestamp(temporal_source, temporal_value))) {
-    return temporal_value;
-  }
-  if (coercion_kind == XmlaCoercionKind::TIME &&
-      (TryParseTimeValue(temporal_source, temporal_value) ||
-       TryParseTimeFromBaseDateTimestamp(temporal_source, temporal_value))) {
-    return temporal_value;
-  }
-  if ((coercion_kind == XmlaCoercionKind::TIMESTAMP ||
-       coercion_kind == XmlaCoercionKind::TIMESTAMP_TZ) &&
-      TryParseTimestampValue(temporal_source, temporal_value)) {
-    return temporal_value;
-  }
-
-  return Value(std::string(body));
-}
-
-static LogicalType InferLogicalType(const Value &value) {
-  if (value.IsNull()) {
-    return LogicalType::SQLNULL;
-  }
-  switch (value.type().id()) {
-  case LogicalTypeId::BOOLEAN:
-    return LogicalType::BOOLEAN;
-  case LogicalTypeId::BIGINT:
-    return LogicalType::BIGINT;
-  case LogicalTypeId::UBIGINT:
-    return LogicalType::UBIGINT;
-  case LogicalTypeId::DOUBLE:
-    return LogicalType::DOUBLE;
-  case LogicalTypeId::DATE:
-    return LogicalType::DATE;
-  case LogicalTypeId::TIME:
-    return LogicalType::TIME;
-  case LogicalTypeId::TIMESTAMP:
-    return LogicalType::TIMESTAMP;
-  case LogicalTypeId::TIMESTAMP_TZ:
-    return LogicalType::TIMESTAMP_TZ;
-  default:
-    return LogicalType::VARCHAR;
-  }
-}
-
-static LogicalType MergeLogicalTypes(const LogicalType &current,
-                                     const LogicalType &next) {
-  if (current.id() == LogicalTypeId::SQLNULL) {
-    return next;
-  }
-  if (next.id() == LogicalTypeId::SQLNULL) {
-    return current;
-  }
-  if (current == next) {
-    return current;
-  }
-
-  auto current_id = current.id();
-  auto next_id = next.id();
-  auto current_numeric = current_id == LogicalTypeId::BIGINT ||
-                         current_id == LogicalTypeId::UBIGINT ||
-                         current_id == LogicalTypeId::DOUBLE ||
-                         current_id == LogicalTypeId::HUGEINT;
-  auto next_numeric =
-      next_id == LogicalTypeId::BIGINT || next_id == LogicalTypeId::UBIGINT ||
-      next_id == LogicalTypeId::DOUBLE || next_id == LogicalTypeId::HUGEINT;
-  if (current_numeric && next_numeric) {
-    if (current_id == LogicalTypeId::DOUBLE ||
-        next_id == LogicalTypeId::DOUBLE) {
-      return LogicalType::DOUBLE;
-    }
-    if (current_id == LogicalTypeId::HUGEINT ||
-        next_id == LogicalTypeId::HUGEINT) {
-      return LogicalType::HUGEINT;
-    }
-    if ((current_id == LogicalTypeId::BIGINT &&
-         next_id == LogicalTypeId::UBIGINT) ||
-        (current_id == LogicalTypeId::UBIGINT &&
-         next_id == LogicalTypeId::BIGINT)) {
-      return LogicalType::HUGEINT;
-    }
-    if (current_id == LogicalTypeId::UBIGINT &&
-        next_id == LogicalTypeId::UBIGINT) {
-      return LogicalType::UBIGINT;
-    }
-    return LogicalType::BIGINT;
-  }
-
-  if ((current_id == LogicalTypeId::DATE &&
-       next_id == LogicalTypeId::TIMESTAMP) ||
-      (current_id == LogicalTypeId::TIMESTAMP &&
-       next_id == LogicalTypeId::DATE)) {
-    return LogicalType::TIMESTAMP;
-  }
-  if ((current_id == LogicalTypeId::DATE &&
-       next_id == LogicalTypeId::TIMESTAMP_TZ) ||
-      (current_id == LogicalTypeId::TIMESTAMP_TZ &&
-       next_id == LogicalTypeId::DATE) ||
-      (current_id == LogicalTypeId::TIMESTAMP &&
-       next_id == LogicalTypeId::TIMESTAMP_TZ) ||
-      (current_id == LogicalTypeId::TIMESTAMP_TZ &&
-       next_id == LogicalTypeId::TIMESTAMP)) {
-    return LogicalType::TIMESTAMP_TZ;
-  }
-
-  return LogicalType::VARCHAR;
 }
 
 struct XmlTag {
@@ -3681,24 +2820,6 @@ static bool TryParseSsasRowsFast(
   return false;
 }
 
-static bool EnableSsasFastRowParser() {
-  auto *raw_value = std::getenv("PBI_SCANNER_ENABLE_SSAS_FAST_ROWS");
-  if (!raw_value || !*raw_value) {
-    return false;
-  }
-  auto value = StringUtil::Lower(Trimmed(raw_value));
-  return value == "1" || value == "true" || value == "yes" || value == "on";
-}
-
-static bool EnableStreamingSxParser() {
-  auto *raw_value = std::getenv("PBI_SCANNER_ENABLE_STREAMING_SX");
-  if (!raw_value || !*raw_value) {
-    return false;
-  }
-  auto value = StringUtil::Lower(Trimmed(raw_value));
-  return value == "1" || value == "true" || value == "yes" || value == "on";
-}
-
 static bool IsRecoverableEarlyFramingError(const Exception &ex,
                                            idx_t candidate_offset) {
   if (candidate_offset >= 8) {
@@ -3968,18 +3089,7 @@ Value CoerceXmlValueForTesting(const std::string &raw_value,
 std::string
 EffectiveExecutionTransportForTesting(const std::string &statement) {
   (void)statement;
-  auto mode = ResolveXmlaTransportMode();
-  switch (mode) {
-  case XmlaTransportMode::PLAIN:
-    return "plain";
-  case XmlaTransportMode::SX:
-    return "sx";
-  case XmlaTransportMode::XPRESS:
-    return "xpress";
-  case XmlaTransportMode::SX_XPRESS:
-    return "sx_xpress";
-  }
-  return "plain";
+  return XmlaTransportName(ResolveXmlaTransportMode());
 }
 
 std::string BuildXmlaExecuteEnvelope(const std::string &catalog,
